+++ apps/client/src/widgets/workspace/WorkspaceContent.tsx
@@ -1,9 +1,14 @@
 import 'blockly/blocks';
 import * as Blockly from 'blockly/core';
 import { useEffect, useState } from 'react';
-import htmlCodeGenerator from '@/widgets/workspace/htmlCodeGenerator';
-import CustomCategory from './customCategory';
+import htmlCodeGenerator from '@/widgets/workspace/blockly/htmlCodeGenerator';
+import CustomCategory from '@/widgets/workspace/blockly/customCategory';
 import { CssPropsSelectBox } from './CssPropsSelectBox';
+import { defineBlocks } from '@/widgets/workspace/blockly/defineBlocks';
+import { toolboxConfig } from '@/widgets/workspace/blockly/toolboxConfig';
+import { initTheme } from '@/widgets/workspace/blockly/initTheme';
+import { customizeFlyoutSVG } from '@/widgets/workspace/blockly/customizeFlyoutSVG';
+import { classMakerPrompt } from './blockly/classMakerPrompt';
 
 Blockly.registry.register(
   Blockly.registry.Type.TOOLBOX_ITEM,
@@ -12,214 +17,19 @@ Blockly.registry.register(
   true
 );
 
-const customTheme = Blockly.Theme.defineTheme('custom', {
-  name: 'custom',
-  base: Blockly.Themes.Classic,
-  componentStyles: {
-    workspaceBackgroundColour: '#fafafa', // 워크스페이스 배경색
-    toolboxBackgroundColour: 'blackBackground', // 툴박스 배경색
-    flyoutBackgroundColour: 'white', // 툴박스 플라이아웃 배경색
-    flyoutOpacity: 1,
-    scrollbarColour: '#000000',
-    insertionMarkerColour: '#fff',
-    insertionMarkerOpacity: 0.3,
-    scrollbarOpacity: 0.001,
-    cursorColour: '#d0d0d0',
-  },
-
-  categoryStyles: {
-    containerCategory: {
-      colour: 'FF3A61',
-    },
-    textCategory: {
-      colour: 'FFD900',
-    },
-    formCategory: {
-      colour: 'FF9821',
-    },
-    tableCategory: {
-      colour: 'B223F5',
-    },
-    listCategory: {
-      colour: '3ED5FF',
-    },
-    linkCategory: {
-      colour: '3E84FF',
-    },
-    etcCategory: {
-      colour: '00AF6F',
-    },
-  },
-});
-
-Blockly.Blocks['html'] = {
-  init: function () {
-    this.appendDummyInput().appendField('html');
-    this.appendValueInput('css class').setCheck('String').appendField('css class');
-    this.appendStatementInput('children').appendField('children');
-    this.setColour(230);
-  },
-};
-
-Blockly.Blocks['head'] = {
-  init: function () {
-    this.setPreviousStatement(true);
-    this.setNextStatement(true);
-    this.appendEndRowInput().appendField('head');
-    this.appendValueInput('css class').setCheck('CSS-CLASS').appendField('css class');
-    this.appendStatementInput('children').appendField();
-    this.setColour(120);
-  },
-};
-
-Blockly.Blocks['body'] = {
-  init: function () {
-    this.setPreviousStatement(true);
-    this.setNextStatement(true);
-    this.appendEndRowInput().appendField('body');
-    this.appendValueInput('css class').setCheck('CSS-CLASS').appendField('css class');
-    this.appendStatementInput('children').appendField();
-    this.setColour(300);
-  },
-};
-
-Blockly.Blocks['p'] = {
-  init: function () {
-    this.setPreviousStatement(true);
-    this.setNextStatement(true);
-    this.appendEndRowInput().appendField('p');
-    this.appendValueInput('css class').setCheck('CSS-CLASS').appendField('css class');
-    this.appendStatementInput('children').appendField();
-    this.setColour(180);
-  },
-};
-
-Blockly.Blocks['button'] = {
-  init: function () {
-    this.setPreviousStatement(true);
-    this.setNextStatement(true);
-    this.appendEndRowInput().appendField('button');
-    this.appendValueInput('css class').setCheck('CSS-CLASS').appendField('css class');
-    this.appendStatementInput('children').appendField();
-    this.setColour(280);
-  },
-};
-
-Blockly.Blocks['text'] = {
-  init: function () {
-    this.setPreviousStatement(true); // 다른 블록 위에 연결 가능
-    this.setNextStatement(true); // 다른 블록 아래에 연결 가능
-    this.appendDummyInput().appendField('text').appendField(new Blockly.FieldTextInput(), 'TEXT');
-    this.setColour(40);
-  },
-};
-
-// css 블록
-Blockly.Blocks['css_style'] = {
-  init: function () {
-    this.appendDummyInput().appendField(new Blockly.FieldTextInput('클래스명'), 'CLASS'); // "클래스명"은 초기값
-    this.setOutput(true); // 이 블록을 다른 블록에 연결할 수 있도록 설정
-  },
-};
-
-const contents = [
-  {
-    kind: 'block',
-    type: 'html',
-  },
-  {
-    kind: 'block',
-    type: 'head',
-  },
-  {
-    kind: 'block',
-    type: 'body',
-  },
-  {
-    kind: 'block',
-    type: 'p',
-  },
-  {
-    kind: 'block',
-    type: 'button',
-  },
-  {
-    kind: 'block',
-    type: 'text',
-  },
-];
-
-const toolboxConfig = {
-  kind: 'categoryToolbox',
-  contents: [
-    {
-      kind: 'category',
-      name: '컨테이너',
-      categorystyle: 'containerCategory',
-      contents: contents,
-    },
-    {
-      kind: 'category',
-      name: '텍스트',
-      categorystyle: 'textCategory',
-      contents: contents,
-    },
-    {
-      kind: 'category',
-      name: '폼',
-      categorystyle: 'formCategory',
-      contents: [
-        {
-          kind: 'button',
-          text: '추가하기',
-          callbackKey: 'openTypedVariableModal',
-        },
-        { kind: 'block', type: 'css_style' },
-      ],
-    },
-    {
-      kind: 'category',
-      name: '표',
-      categorystyle: 'tableCategory',
-      contents: contents,
-    },
-    {
-      kind: 'category',
-      name: '리스트',
-      categorystyle: 'listCategory',
-      contents: contents,
-    },
-    {
-      kind: 'category',
-      name: '링크',
-      categorystyle: 'linkCategory',
-      contents: contents,
-    },
-    {
-      kind: 'category',
-      name: '기타',
-      categorystyle: 'etcCategory',
-      contents: contents,
-    },
-  ],
-};
-
-const toolboxConfig2 = {
-  kind: 'categoryToolbox',
-  contents: [],
-};
-
 export const WorkspaceContent = () => {
   const [workspace, setWorkspace] = useState<Blockly.WorkspaceSvg | null>(null);
   const [htmlCode, setHtmlCode] = useState<string>('');
   const [activeTab, setActiveTab] = useState<'preview' | 'html' | 'css'>('preview');
+
+  defineBlocks();
 
   useEffect(() => {
     const newWorkspace = Blockly.inject('blocklyDiv', {
       renderer: 'zelos',
       toolboxPosition: 'end',
       toolbox: toolboxConfig,
-      theme: customTheme,
+      theme: initTheme,
       zoom: {
         controls: true,
         wheel: true,
@@ -230,98 +40,15 @@ export const WorkspaceContent = () => {
       },
     });
 
-    newWorkspace.registerButtonCallback('openTypedVariableModal', () =>
-      openTypedVariableModal(newWorkspace)
-    );
+    newWorkspace.registerButtonCallback('classMakerPrompt', () => classMakerPrompt(newWorkspace));
 
     setWorkspace(newWorkspace);
-
-    interface IExtendedIToolbox extends Blockly.IToolbox {
-      HtmlDiv: HTMLElement;
-    }
-
-    const customizeFlyoutSVG = () => {
-      const toolbox: IExtendedIToolbox = newWorkspace.getToolbox()! as IExtendedIToolbox;
-
-      const tabs = document.createElement('div');
-      tabs.className = 'flex w-96';
-
-      const tab1 = document.createElement('button');
-      tab1.classList.add('tab');
-      tab1.textContent = 'HTML';
-
-      const tab2 = document.createElement('button');
-      tab2.classList.add('tab');
-      tab2.textContent = 'CSS';
-
-      tab1.addEventListener('click', () => {
-        newWorkspace.updateToolbox(toolboxConfig);
-        const toolboxContents = document.querySelector('.blocklyToolboxContents');
-        toolboxContents!.classList.remove('hidden');
-        tab1.classList.add('tabSelected');
-        tab2.classList.remove('tabSelected');
-      });
-
-      tab2.addEventListener('click', () => {
-        newWorkspace.updateToolbox(toolboxConfig2);
-        const toolboxContents = document.querySelector('.blocklyToolboxContents');
-        toolboxContents!.classList.add('hidden');
-        tab2.classList.add('tabSelected');
-        tab1.classList.remove('tabSelected');
-      });
-
-      tabs.appendChild(tab1);
-      tabs.appendChild(tab2);
-
-      toolbox!.HtmlDiv.prepend(tabs);
-      const flyout = newWorkspace!.getToolbox()!.getFlyout();
-      flyout!.hide = () => {};
-    };
-
-    customizeFlyoutSVG();
+    customizeFlyoutSVG(newWorkspace);
 
     return () => {
       newWorkspace.dispose();
     };
   }, []);
-
-  // prompt를 이용한 동적 생성
-  const openTypedVariableModal = (workspace: Blockly.WorkspaceSvg) => {
-    const blockName = prompt('새로운 css 블록 이름을 입력하세요.');
-
-    if (blockName?.trim() === '') {
-      return alert('블록 이름을 입력해주세요.');
-    }
-
-    if (!Blockly.Blocks[blockName!]) {
-      Blockly.Blocks[blockName!] = {
-        init: function () {
-          this.appendDummyInput().appendField(new Blockly.FieldTextInput(blockName!), 'CLASS'); // 입력된 이름 반영
-          this.setOutput(true);
-          this.setColour(230);
-        },
-      };
-    }
-
-    // "폼" 카테고리를 찾아 기존 블록 유지 및 새 블록 추가
-    const formCategory = toolboxConfig.contents.find((category) => category.name === '폼');
-
-    // 기존 블록 유지 및 새 블록 추가
-    const existingBlocks = formCategory!.contents || [];
-    const isBlockAlreadyAdded = existingBlocks.some((block) => block.type === blockName);
-
-    if (isBlockAlreadyAdded) {
-      alert(`"${blockName}" 블록은 이미 "폼" 카테고리에 존재합니다.`);
-      return;
-    }
-
-    formCategory!.contents = [...existingBlocks, { kind: 'block', type: blockName }];
-
-    // 기존 툴박스 갱신
-    workspace.updateToolbox(toolboxConfig);
-
-    alert(`새 블록 "${blockName}"이(가) "폼" 카테고리에 성공적으로 추가되었습니다.`);
-  };
 
   const generateHtmlCode = () => {
     if (!workspace) {
