+++ apps/client/src/widgets/workspace/blockly/categoryStyles.ts
@@ -1,0 +1,23 @@
+export const categoryColours = {
+  containerCategory: {
+    colour: 'FF3A61',
+  },
+  textCategory: {
+    colour: 'FFD900',
+  },
+  formCategory: {
+    colour: 'FF9821',
+  },
+  tableCategory: {
+    colour: 'B223F5',
+  },
+  listCategory: {
+    colour: '3ED5FF',
+  },
+  linkCategory: {
+    colour: '3E84FF',
+  },
+  etcCategory: {
+    colour: '00AF6F',
+  },
+};
